cleanup parent interface handling
